Gracefully handle parser errors
Replace the default error listener with a custom one that throws a ParserError
exception when one occurs. This way we can differentiate any old exception
occuring in parsing from a parser error. This also means that parser errors
don't crash our code, which is generally a good thing. Eventually we may be
able to determine what precisely was mis-parsed.