Added new tests and a general GrapheneStore class
Made some changes to the Node class so that the inUse byte is stored
as a boolean (was causing the failure of some tests). Also added
overloading for the == operator for the Node class. Added a new test
that writes two nodes to the file and checks if they are read as
written. Will need to test overwriting next. Added a general
GrapheneStore class to handle tasks that all the stores need to
perform, currently this is just finding the directory where db files
will be stored. It also has a class variable TESTING that, if set,
will change the directory of the db files from "datafiles" to
"test_datafiles". Also added padding for the first 9 bytes of the
nodestore file, this allows for error checking in node creation if
attempting to write the node to index 0.