Parser errors now point to offending token
Change whitespace/comment parsing rules so that they don't skip the matches,
but hide them so that we can still see the user's original input.

On parser error, message points to the token that broke the parsing mechanism.
Soon we may be able to detect more specifically what a user meant.