Set up build directory
When running tests (with coverage) and/or making docs, files will now go into
build directory.

There is an index file for the build directory that resides in
res/build-index.html, which is copied over every time the build directory is
used.

One can view the build directory easily with `make build-server`, then going to
http://localhost:12323/.