Finished implementation of Name and NameStore classes and tests
Will now work on index recycling and a higher level manager that
communicates between the store classes. Also made a change to the
way IDs are stored. Changed them from unsigned ints to signed ints
to allow for storage of negative values that might signify an
error (i.e. -1 for there being no previous relationship in the
doubly linked list). Will have to determine how to handle the
PropertyStore's propBlockID since it needs to be able to store
both unsigned and signed number values, but struct will throw an
exception if the value is signed when it should be unsigned.
Perhaps creating a different STRUCT_FORMAT_STR for each of the
different types will solve this problem.