Finished implementation of GeneralStore class
A lot of code was removed and refactored with the creation of the GeneralStore
class. Tests were kept back in order to focus on funcitonality, but a lot of
them can be removed, and only tested in the GeneralStore since that is where
most of the code lies. Also finished implementation of the RelationshipType
and RelationshipType classes, as well as testing for those classes.