Changed GeneralStore implementation and fixed the PropertyStore
Changed GeneralStore implementation to use the RECORD_SIZE instead of the
structure of the item (it was unneccessary information since it only read
and wrote according to the record size of the given struct). Also fixed the
problem of the PropertyStore not being able to store actual values other than
unsigned numbers correctly, now all primitives are supported, and the dynamic
types should be handled by the dynamic stores.