Fix issue with deleting types and recreating them
When types were deleted, then recreated, the datafiles were zeroed out instead
of deleted, so when adding in new nodes, the old value of None would still be
around. But we used None for both a deleted value and for an EOF demarcation,
so a separate EOF demarcation was added.