Many changes to the StorageManager and its tests
StorageManager is now more abstracted to handle relationship
commands, it also now distinguishes between node and relationship
types and type types. Also fixed the assert statements in the test
file to use AssertEquals instead, as well as to handle the flag
used to denote whether the command is for a relationship or a node;
will abstract this into different methods shortly.