Fix errors in NodePropertyStore and StorageManager
NodePropertyStore needs to throw an error when deleting an EOF key, since you
cannot access the return value of a deletion. Also it wouldn't make sense to
return something there anyway, since trying to delete out of bounds is an
error.

StorageManager also broke when no properties existed in a Type.