MATCH command now handles relations.
Relations are handled by creating a tree where the right leaf is always
a NodeIterator instance and the left leaf is either a NodeIterator or
a RelationIterator. RelationIterators can take a NI or an RI as a left child.
We can handle arbitrary degrees of separation.

Still does not handle filtering out columns that were not specifically asked
for.