Finished tests for Array and ArrayStore
Finished adding unittests for the array type and the ArrayStore.
Notice a new method almost_equal was added to the Array type, this
is to handle testing for float types stored which are imprecise;
it is recommended to store decimals as doubles to avoid this imprecision.
For querying, an almost equal operator (~=) will probably need to be
added in order to query decimal types.