Truncation Changes
Made it so the file is truncated when it has several IDs that are free at
the end of the file. Made an implementation change: EOF is no longer a
hard-failure (raise exception) since EOFs may occur because of file truncation.
They should not normally occur, which is why the EOF exception was replaced
with a logger.warn message about unexpected EOFs. The tests were changed
accordingly to reflect this change. Now EOFs and None returns are treated
equally since an EOF is the same as reading an item which was deleted from
the end of the file which would have otherwise returned None, but now
returns EOF. Loggers were added to ArrayManager and NameManager.