Fix cascading deletes of relations
When nodes are deleted, the relations that refer to them must be deleted too.
However, the functions were using the cache to get the node, which loaded
properties from the node; since the properties were being deleted in the
delete_node function BEFORE the relations were, this threw an error every time.